Change autoBounce to bounceOnAdd, default to false
The name of the autoBounce option has changed to bounceOnAdd as it reflects more
on what will happen if you use it.

Also, there was not a lot of reasons to default it to true so when you added
a Marker it would bounce.
Now it default to false, if you want your marker to bounce when added, you must
specify it.