Fix bounce() bounce at the wrong position
The animation is fine on map load, but the markers are bouncing at the
wrong place when the user has zoomed or dragged the map and one is
calling the "bounce" method on the markers afterwards.

To get the marker position on the screen (x, y), we should use the
original latitude and longitude (don't change) and not the original
pixel coordinates (change upon zoom/drag).

Rework the commit log. Maxime Hadjinlian <maxime.hadjinlian@gmail.com>