bouncemarker.js: use L.Utils.requestAnimFrame
Instead of disrupting the browser flow by using setInterval to do
animation (especially because behing the scenes Leaflet is already using
CSS transform and requestAnimationFrame()); use the
L.Utils.requestAnimFrame.

This makes the plugin way gentler on the CPU and the animation way
smoother.
After testing it with a few (~100) points, it makes a huge difference,
especially on mobile.