bouncemarker.js: remove public function
There's no need for the public function anymore, the only interest was
that it would call getLatLng(); but it's already called by the onAdd()
function and you need to add the Marker to a map to use it.

So it doesn't make sense and it generates bug.

So removing this line, the public function is simply a mailbox, so let's
remove it altogether.